Refactor the macros
- remove redundancy in reading and writing macros between Document and Element
- Pass in expected values as identifiers rather than relying on having
  consistent names across functions
